requirements.txt
fastapi~=0.103.2
SQLAlchemy[asyncio]~=2.0.22
starlette~=0.27.0
python-dotenv~=1.0.0
h11~=0.14.0
pip~=23.3.1
Mako~=1.2.4
MarkupSafe~=2.1.3
PyYAML~=6.0.1
anyio~=3.7.1
sniffio~=1.3.0
click~=8.1.7
alembic~=1.12.1
asyncpg~=0.28.0
pydantic[email]~=2.4.2
websockets~=11.0.3
httptools~=0.6.0
watchfiles~=0.21.0
greenlet~=3.0.1
uvicorn
gunicorn
aioftp~=0.21.4
aiofiles~=23.2.1
fastapi-cache2[redis]
redis~=4.6.0
httpx
pytest-asyncio
fakeredis
fastapi-users[sqlalchemy,oauth]
pyotp